[testing] Refactor
